Move Checks test to a Checks directory and namespace
