+++ tests/Checks/AppKeyIsSetTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BeyondCode\SelfDiagnosis\Tests;
+namespace BeyondCode\SelfDiagnosis\Tests\Checks;
 
 use Orchestra\Testbench\TestCase;
 use BeyondCode\SelfDiagnosis\Checks\AppKeyIsSet;
