+++ tests/Checks/CorrectPhpVersionIsInstalledTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BeyondCode\SelfDiagnosis\Tests;
+namespace BeyondCode\SelfDiagnosis\Tests\Checks;
 
 use Orchestra\Testbench\TestCase;
 use BeyondCode\SelfDiagnosis\Checks\CorrectPhpVersionIsInstalled;
@@ -10,7 +10,7 @@ class CorrectPhpVersionIsInstalledTest extends TestCase
     /** @test */
     public function it_detects_the_php_version_from_composer_json()
     {
-        $this->app->setBasePath(__DIR__ . '/fixtures');
+        $this->app->setBasePath(__DIR__ . '/../fixtures');
 
         $check = app(CorrectPhpVersionIsInstalled::class);
