+++ tests/Checks/DatabaseCanBeAccessedTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BeyondCode\SelfDiagnosis\Tests;
+namespace BeyondCode\SelfDiagnosis\Tests\Checks;
 
 use Illuminate\Support\Facades\DB;
 use Orchestra\Testbench\TestCase;
