+++ tests/Checks/MigrationsAreUpToDateTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BeyondCode\SelfDiagnosis\Tests;
+namespace BeyondCode\SelfDiagnosis\Tests\Checks;
 
 use Illuminate\Support\Facades\Artisan;
 use Orchestra\Testbench\TestCase;
