+++ tests/Checks/StorageDirectoryIsLinkedTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BeyondCode\SelfDiagnosis\Tests;
+namespace BeyondCode\SelfDiagnosis\Tests\Checks;
 
 use BeyondCode\SelfDiagnosis\Checks\StorageDirectoryIsLinked;
 use Illuminate\Filesystem\Filesystem;
